# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soltex Global Admin API",
    "version": "0.3.0",
    "description": "Phase C: authentication, second factors, users and audit. Errors are RFC 9457 Problem Details (application/problem+json) with `request_id`; every response carries `X-Request-ID`. State-changing requests need the header `x-requested-with: soltex-admin`."
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-soltex_session",
        "description": "Opaque session token (HttpOnly cookie set by /auth/login)."
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-requested-with",
        "description": "Must be \"soltex-admin\" on POST/PUT/PATCH/DELETE."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ],
              "additionalProperties": false
            }
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "request_id"
        ],
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details"
      },
      "SessionView": {
        "type": "object",
        "properties": {
          "next": {
            "type": "string",
            "enum": [
              "authenticated",
              "mfa_required",
              "mfa_enrollment_required"
            ]
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "email": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "owner",
                  "admin",
                  "content_manager",
                  "editor",
                  "seo_specialist",
                  "marketer"
                ]
              }
            },
            "required": [
              "id",
              "email",
              "name",
              "role"
            ],
            "additionalProperties": false
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "content.read",
                "content.edit",
                "content.publish",
                "seo.read",
                "seo.edit",
                "seo.publish",
                "redirects.read",
                "redirects.manage",
                "media.read",
                "media.upload",
                "media.replace",
                "media.delete",
                "leads.read",
                "leads.manage",
                "leads.export",
                "users.manage",
                "settings.manage",
                "system.read",
                "system.deploy",
                "system.rollback",
                "audit.read"
              ]
            }
          },
          "session": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "aal": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "expiresAt": {
                "type": "string"
              },
              "stepUpValidUntil": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "id",
              "aal",
              "expiresAt",
              "stepUpValidUntil"
            ],
            "additionalProperties": false
          },
          "mfa": {
            "type": "object",
            "properties": {
              "requirement": {
                "type": "string",
                "enum": [
                  "passkey",
                  "any",
                  "none"
                ]
              },
              "requirementMet": {
                "type": "boolean"
              },
              "enrolled": {
                "type": "boolean"
              },
              "passkeys": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "totp": {
                "type": "boolean"
              },
              "recoveryCodesLeft": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "requirement",
              "requirementMet",
              "enrolled",
              "passkeys",
              "totp",
              "recoveryCodesLeft"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "next",
          "user",
          "permissions",
          "session",
          "mfa"
        ],
        "additionalProperties": false
      },
      "RecoveryCodes": {
        "type": "object",
        "properties": {
          "recoveryCodes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Shown once. Store them offline."
          }
        },
        "required": [
          "recoveryCodes"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/mfa": {
      "get": {
        "summary": "My second factors and the requirement of my role",
        "tags": [
          "mfa"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "x-soltex-access": {
          "level": "enrollment"
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/totp/verify": {
      "post": {
        "summary": "Verify a TOTP code (sign-in second factor or step-up)",
        "tags": [
          "mfa"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "code"
                ]
              }
            }
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "pending"
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/recovery/verify": {
      "post": {
        "summary": "Use a one-time recovery code (lost device)",
        "tags": [
          "mfa"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 64
                  }
                },
                "required": [
                  "code"
                ]
              }
            }
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "pending"
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/totp/enroll": {
      "post": {
        "summary": "Start TOTP enrollment",
        "tags": [
          "mfa"
        ],
        "description": "Returns the secret and the otpauth:// URI (for a QR code) once. Confirm with a code to activate.",
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "enrollment",
          "stepUpIfEnrolled": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": {
                      "type": "string"
                    },
                    "otpauthUri": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "secret",
                    "otpauthUri"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/totp/confirm": {
      "post": {
        "summary": "Confirm TOTP enrollment with a first code",
        "tags": [
          "mfa"
        ],
        "description": "If this is the first second factor, the session is raised to assurance 2 and recovery codes are returned (once).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "code"
                ]
              }
            }
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "enrollment",
          "stepUpIfEnrolled": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "next": {
                      "type": "string",
                      "enum": [
                        "authenticated",
                        "mfa_required",
                        "mfa_enrollment_required"
                      ]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "owner",
                            "admin",
                            "content_manager",
                            "editor",
                            "seo_specialist",
                            "marketer"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "name",
                        "role"
                      ],
                      "additionalProperties": false
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "content.read",
                          "content.edit",
                          "content.publish",
                          "seo.read",
                          "seo.edit",
                          "seo.publish",
                          "redirects.read",
                          "redirects.manage",
                          "media.read",
                          "media.upload",
                          "media.replace",
                          "media.delete",
                          "leads.read",
                          "leads.manage",
                          "leads.export",
                          "users.manage",
                          "settings.manage",
                          "system.read",
                          "system.deploy",
                          "system.rollback",
                          "audit.read"
                        ]
                      }
                    },
                    "session": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "aal": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "expiresAt": {
                          "type": "string"
                        },
                        "stepUpValidUntil": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "aal",
                        "expiresAt",
                        "stepUpValidUntil"
                      ],
                      "additionalProperties": false
                    },
                    "mfa": {
                      "type": "object",
                      "properties": {
                        "requirement": {
                          "type": "string",
                          "enum": [
                            "passkey",
                            "any",
                            "none"
                          ]
                        },
                        "requirementMet": {
                          "type": "boolean"
                        },
                        "enrolled": {
                          "type": "boolean"
                        },
                        "passkeys": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "totp": {
                          "type": "boolean"
                        },
                        "recoveryCodesLeft": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "requirement",
                        "requirementMet",
                        "enrolled",
                        "passkeys",
                        "totp",
                        "recoveryCodesLeft"
                      ],
                      "additionalProperties": false
                    },
                    "recoveryCodes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "next",
                    "user",
                    "permissions",
                    "session",
                    "mfa"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/totp": {
      "delete": {
        "summary": "Disable TOTP (step-up required)",
        "tags": [
          "mfa"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "full",
          "stepUp": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/recovery-codes/regenerate": {
      "post": {
        "summary": "Generate new recovery codes (step-up required)",
        "tags": [
          "mfa"
        ],
        "description": "All previous codes stop working. The new codes are returned once.",
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "full",
          "stepUp": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "security": [
    {
      "session": [],
      "csrf": []
    }
  ],
...
}